Move 'movement' out of grid module
