+++ src/game/ghost.rs
@@ -4,14 +4,11 @@ use crate::{
     game::{animation::PlayerAnimation, assets::ImageAsset},
     screen::Screen,
 };
-
+use crate::game::movement::{GridMovement, Roll};
 use super::{
     assets::ImageAssets,
     end_game::EndGameCondition,
-    grid::{
-        movement::{GridMovement, Roll},
-        GridPosition,
-    },
+    grid::GridPosition,
     line_of_sight::LineOfSightBundle,
     spawn::{
         health::{OnDeath, SpawnPointGridPosition},
