+++ src/game/mod.rs
@@ -11,10 +11,10 @@ pub mod end_game;
 mod ghost;
 pub mod grid;
 pub mod line_of_sight;
-mod movement;
 pub mod spawn;
 pub mod stamina;
 mod threat;
+pub mod movement;
 
 pub(super) fn plugin(app: &mut App) {
     app.add_plugins((
