+++ src/game/movement.rs
@@ -1,77 +1,150 @@
-//! Handle player input and translate it into movement.
-//! Note that the approach used here is simple for demonstration purposes.
-//! If you want to move the player in a smoother way,
-//! consider using a [fixed timestep](https://github.com/bevyengine/bevy/blob/latest/examples/movement/physics_in_fixed_timestep.rs).
-
-use bevy::prelude::*;
-
-use crate::game::spawn::player::Player;
+/// Grid-based movement
+use crate::input::PlayerAction;
 use crate::AppSet;
+use bevy::prelude::*;
+use leafwing_input_manager::prelude::ActionState;
+use std::time::Duration;
+use crate::game::grid::GridPosition;
 
-pub(super) fn plugin(app: &mut App) {
-    // Record directional input as movement controls.
-    app.register_type::<MovementController>();
+pub fn plugin(app: &mut App) {
+    app.add_systems(Update, update_roll_timer.in_set(AppSet::TickTimers));
     app.add_systems(
         Update,
-        record_movement_controller.in_set(AppSet::RecordInput),
+        (respond_to_input, apply_movement)
+            .chain()
+            .in_set(AppSet::UpdateVirtualGrid),
     );
 
-    // Apply movement based on controls.
-    app.add_systems(Update, apply_movement.chain().in_set(AppSet::Update));
+
+    app.register_type::<GridMovement>();
 }
 
-#[derive(Component, Reflect, Default)]
+#[derive(Component, Reflect, Debug, Copy, Clone, PartialEq)]
 #[reflect(Component)]
-pub struct MovementController(pub Vec2);
-
-#[derive(Event, Reflect)]
-pub struct UseStamina;
-
-fn record_movement_controller(
-    input: Res<ButtonInput<KeyCode>>,
-    mut controller_query: Query<&mut MovementController, With<Player>>,
-) {
-    // Collect directional input.
-    let mut intent = Vec2::ZERO;
-    if input.pressed(KeyCode::KeyW) || input.pressed(KeyCode::ArrowUp) {
-        intent.y += 1.0;
-    }
-    if input.pressed(KeyCode::KeyS) || input.pressed(KeyCode::ArrowDown) {
-        intent.y -= 1.0;
-    }
-    if input.pressed(KeyCode::KeyA) || input.pressed(KeyCode::ArrowLeft) {
-        intent.x -= 1.0;
-    }
-    if input.pressed(KeyCode::KeyD) || input.pressed(KeyCode::ArrowRight) {
-        intent.x += 1.0;
-    }
-
-    // Normalize so that diagonal movement has the same speed as
-    // horizontal and vertical movement.
-    let intent = intent.normalize_or_zero();
-
-    // Apply movement intent to controllers.
-    for mut controller in &mut controller_query {
-        controller.0 = intent;
-    }
+pub struct GridMovement {
+    pub velocity: Vec2,
+    pub friction: f32,
+    pub acceleration_player_force: Vec2,
+    pub acceleration_external_force: Vec2,
+    pub acceleration_player_multiplier: f32,
+    pub is_rolling: bool,
 }
 
-#[derive(Component, Reflect, Default)]
+#[derive(Component, Reflect, Debug, PartialEq)]
 #[reflect(Component)]
-pub struct Movement {
-    /// Since Bevy's default 2D camera setup is scaled such that
-    /// one unit is one pixel, you can think of this as
-    /// "How many pixels per second should the player move?"
-    /// Note that physics engines may use different unit/pixel ratios.
-    pub speed: f32,
+pub struct Roll {
+    pub timer: Timer,
+    pub velocity_multiplier: f32,
 }
 
-fn apply_movement(
+impl Roll {
+    pub const ROLL_TIME: Duration = Duration::from_millis(300);
+}
+
+impl Default for Roll {
+    fn default() -> Self {
+        Self {
+            timer: Timer::from_seconds(0.5, TimerMode::Once),
+            velocity_multiplier: 3.0,
+        }
+    }
+}
+
+impl GridMovement {
+    pub fn immobile() -> Self {
+        Self {
+            velocity: Vec2::ZERO,
+            friction: 0.0,
+            acceleration_player_force: Vec2::ZERO,
+            acceleration_external_force: Vec2::ZERO,
+            acceleration_player_multiplier: 0.,
+            is_rolling: false,
+        }
+    }
+}
+
+impl Default for GridMovement {
+    // todo create "presets" like slow, medium, fast for use by enemies or players
+    fn default() -> Self {
+        Self {
+            velocity: Vec2::ZERO,
+            friction: 0.85,
+            acceleration_player_force: Vec2::ZERO,
+            acceleration_external_force: Vec2::ZERO,
+            acceleration_player_multiplier: 66.,
+            is_rolling: false,
+        }
+    }
+}
+
+impl GridMovement {
+    pub fn current_force(&self) -> Vec2 {
+        self.acceleration_player_force + self.acceleration_external_force
+    }
+}
+
+pub fn respond_to_input(mut query: Query<(&ActionState<PlayerAction>, &mut GridMovement)>) {
+    for (action_state, mut movement) in query.iter_mut() {
+        let mut intent = Vec2::ZERO;
+
+        if action_state.pressed(&PlayerAction::MoveUp) {
+            intent.y += 1.0;
+        }
+        if action_state.pressed(&PlayerAction::MoveDown) {
+            intent.y -= 1.0;
+        }
+        if action_state.pressed(&PlayerAction::MoveLeft) {
+            intent.x -= 1.0;
+        }
+        if action_state.pressed(&PlayerAction::MoveRight) {
+            intent.x += 1.0;
+        }
+        // Normalize so that diagonal movement has the same speed as horizontal and vertical movement.
+        let intent = intent.normalize_or_zero();
+
+        movement.acceleration_player_force = intent * movement.acceleration_player_multiplier;
+
+        if !movement.is_rolling && action_state.pressed(&PlayerAction::Roll) {
+            movement.is_rolling = true;
+        }
+    }
+}
+
+pub fn apply_movement(
+    mut query: Query<(&mut GridPosition, &mut GridMovement, Option<&mut Roll>)>,
     time: Res<Time>,
-    mut movement_query: Query<(&MovementController, &Movement, &mut Transform)>,
 ) {
-    for (controller, movement, mut transform) in &mut movement_query {
-        let velocity = movement.speed * controller.0;
-        transform.translation += velocity.extend(0.0) * (time.delta_seconds() / 8.0);
+    let dt = time.delta_seconds();
+    for (mut position, mut movement, maybe_roll) in query.iter_mut() {
+        let force = movement.current_force() * dt; // scale it by time
+
+        // apply forces and friction
+        let mut velocity = movement.velocity + force;
+        velocity *= movement.friction;
+        if velocity.length() < 0.01 {
+            velocity = Vec2::ZERO;
+        }
+        movement.velocity = velocity;
+
+        // move the player
+        let roll_multi = match (maybe_roll, movement.is_rolling) {
+            (Some(roll), true) => roll.velocity_multiplier,
+            _ => 1.0,
+        };
+
+        position.offset += movement.velocity * dt * roll_multi;
+        position.fix_offset_overflow();
     }
 }
+
+fn update_roll_timer(time: Res<Time>, mut query: Query<(&mut Roll, &mut GridMovement)>) {
+    let dt = time.delta_seconds();
+    for (mut roll, mut movement) in query.iter_mut() {
+        roll.timer.tick(Duration::from_secs_f32(dt));
+        if roll.timer.finished() {
+            movement.is_rolling = false;
+            roll.timer.reset();
+        }
+    }
+}
+
