+++ src/game/spawn/enemy.rs
@@ -1,4 +1,4 @@
-use crate::game::grid::movement::GridMovement;
+use crate::game::movement::GridMovement;
 use crate::game::grid::GridPosition;
 use crate::game::spawn::health::CanApplyDamage;
 use crate::game::spawn::player::Player;
